updated with latest operator loader spec
