+++ awslambda/__init__.py
@@ -1,0 +1,6 @@
+from .delete import delete
+from .deploy import deploy
+from .describe import describe
+from .update import update
+
+__all__ = ["deploy", "describe", "delete", "update"]
