+++ awslambda/describe.py
@@ -1,16 +1,7 @@
-import argparse
-import os
-
 import boto3
 from botocore.exceptions import ClientError
-from rich.pretty import pprint
 
-if __name__ == '__main__':
-    from aws_lambda import generate_lambda_resource_names
-    from utils import get_configuration_value
-else:
-    from .aws_lambda import generate_lambda_resource_names
-    from .utils import get_configuration_value
+from .aws_lambda import generate_lambda_resource_names
 
 
 def describe(deployment_name, lambda_config):
@@ -44,24 +35,3 @@ def describe(deployment_name, lambda_config):
     info_json.update(outputs)
 
     return info_json
-
-
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(
-        description="Describe the bundle deployed on lambda",
-        epilog="Check out https://github.com/bentoml/aws-lambda-deploy#readme to know more",
-    )
-    parser.add_argument(
-        "deployment_name", help="The name you want to use for your deployment"
-    )
-    parser.add_argument(
-        "config_json",
-        help="(optional) The config file for your deployment",
-        default=os.path.join(os.getcwd(), "lambda_config.json"),
-        nargs="?",
-    )
-    args = parser.parse_args()
-
-    lambda_config = get_configuration_value(args.config_json)
-    info_json = describe(args.deployment_name, lambda_config)
-    pprint(info_json)
