+++ operator_config.py
@@ -1,9 +1,7 @@
-from .delete import delete
-from .deploy import deploy
-from .describe import describe
-from .update import update
-
 OPERATOR_NAME = "aws-lambda"
+
+OPERATOR_MODULE = "awslambda"
+
 OPERATOR_SCHEMA = {
     "region": {"required": True, "type": "string", "default": "us-west-1"},
     "timeout": {"required": False, "type": "integer", "coerce": int, "default": 10},
